Cumulated hazard over two time scales

Computes the cumulated hazard surface over two time scales (currently only implemented for objects of class haz2ts, not haz2tsLMM).

Usage

cumhaz2ts(fitted_model, plot_grid = NULL, cause = NULL)

Arguments

fitted_model

The output of the function fit2ts. This is an object of class 'haz2ts'.

plot_grid

(optional) A list containing the parameters to build a new finer grid of intervals over u and s for plotting. This must be of the form:

plot_grid = list(c(umin, umax, du), c(smin, smax, ds)) where umin, umax and smin, smax are the minimum and maximum values desired for the intervals over u and s respectively, and du, ds are distances between intervals over u and s respectively. Specifying a new denser grid is used to evaluate the B-spline bases used for estimation on such grid and plot the estimated surfaces with a greater level of details. If not specified, the plotting is done using the same B-splines bases as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

cause

a character string with a short name for the cause (optional).

Value

A list with the following elements: * Haz the 2d hazard surface (obtained from the function get_hazard_2d); * CumHaz the cumulated hazard estimate over u and s; * cause (if provided) the short name for the cause