Fit a smooth hazard model with one time scale

fit1ts() fits a smooth hazard model with one time scale.

Two methods are implemented for the search of the optimal smoothing parameter (and therefore optimal model): a numerical optimization of the AIC or BIC of the model and a search for the minimum AIC or BIC of the model over a grid of log_10 values for the smoothing parameter. Construction of the B-splines basis and of the penalty matrix is incorporated within the function. If a matrix of covariates is provided, the function will estimate a model with covariates.

Usage

fit1ts(
  data1ts = NULL,
  y = NULL,
  r = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  Wprior = NULL,
  pord = 2,
  optim_method = c("ucminf", "grid_search"),
  optim_criterion = c("aic", "bic"),
  lrho = 0,
  ridge = 0,
  control_algorithm = list(),
  par_gridsearch = list()
)

Arguments

data1ts

(optional) an object created by the function prepare_data(). Providing this input is the easiest way to use the function fit1ts. However, the user can also provide the input data together with a list of bins, as explained by the following parameters' descriptions.

y

A vector of event counts of length ns, or an array of dimension ns by n.

r

A vector of exposure times of length ns, or an array of dimension ns by n.

Z

(optional) A regression matrix of covariates of dimension n by p.

bins

a list with the specification for the bins. This is created by the function prepare_data. Alternatively, a list with the following elements can be provided: * bins_s is a vector of bins extremes for the time scale s. * mids is a vector with the midpoints of the bins over s. * ns is the number of bins over s.

Bbases_spec

A list with the specification for the B-splines basis with the following elements:

bdeg The degree of the B-splines basis. Default is 3 (for cubic B-splines).

nseg_s The number of segments for the B-splines over s. Default is 10.

min_s (optional) The lower limit of the domain of Bs. Default is min(bins_s).

max_s (optional) The upper limit of the domain of Bs. Default is max(bins_s).

Wprior

An optional vector of a-priori weights.

pord

The order of the penalty. Default is 2.

optim_method

The method to be used for optimization: "ucminf" (default) for the numerical optimization of the AIC (or BIC) and "grid_search" for a grid search of the minimum AIC (or BIC) over a grid of log_10(rho_s) values.

optim_criterion

The criterion to be used for optimization: "aic" (default) or "bic".

lrho

A number if optim_method == "ucminf", default is 0. A vector of values for log_10(rho_s) if optim_method == "grid_search". In the latter case, if a vector is not provided, a default sequence of values is used for log_10(rho_s).

ridge

A ridge penalty parameter: default is 0.

control_algorithm

A list with optional values for the parameters of the iterative processes:

maxiter The maximum number of iteration for the IWSL algorithm. Default is 20.

conv_crit The convergence criteria, expressed as difference between estimates at iteration i and i+1. Default is 1e-5.

verbose A Boolean. Default is FALSE. If TRUE monitors the iteration process.

monitor_ev A Boolean. Default is FALSE. If TRUE monitors the evaluation of the model over the log_10(rho_s) values.

par_gridsearch

A list of parameters for the grid_search:

plot_aic A Boolean. Default is FALSE. If TRUE, plot the AIC values over the grid of log_10(rhos) values.

plot_bic A Boolean. Default is FALSE. If TRUE, plot the BIC values over the grid of log_10(rhos) values.

return_aic A Boolean. Default is TRUE. Return the AIC values.

return_bic A Boolean. Default is TRUE. Return the BIC values.

mark_optimal A Boolean. Default is TRUE. If the plot of the AIC or BIC values is returned, marks the optimal log_10(rho_s) in the plot.

main_aic The title of the AIC plot. Default is "AIC grid".

main_bic The title of the BIC plot. Default is "BIC grid".

Value

An object of class haz1ts with the following elements:

optimal_model A list with:

alpha The vector of estimated P-splines coefficients of length cs.

SE_alpha The vector of estimated Standard Errors for the alpha coefficients, of length cs.

beta The vector of estimated covariates coefficients of length p (if model with covariates).

se_beta The vector of estimated Standard Errors for the beta coefficients of length p (if model with covariates).

eta or eta0. The vector of values of the (baseline) linear predictor (log-hazard).

H The hat-matrix.

Cov The full variance-covariance matrix.

deviance The deviance.

ed The effective dimension of the model.

aic The value of the AIC.

bic The value of the BIC.

Bbases a list with the B-spline basis Bs (this is a list for compatibility with functions in 2d).

optimal_logrho The optimal value of log10(rho_s).

P_optimal The optimal penalty matrix P.

AIC (if par_gridsearch$return_aic == TRUE) The vector of AIC values.

BIC (if par_gridsearch$return_bic == TRUE) The vector of BIC values.