Fit a smooth hazard model with two time scales

fit2ts() fits a smooth hazard model with two time scales.

Two methods are implemented for the search of the optimal smoothing parameters (and therefore optimal model): a numerical optimization of the AIC or BIC of the model and a search for the minimum AIC or BIC of the model over a grid of log_10 values for the smoothing parameters. Construction of the B-splines bases and of the penalty matrix is incorporated within the function. If a matrix of covariates is provided, the function will estimate a model with covariates.

Usage

fit2ts(
  data2ts = NULL,
  Y = NULL,
  R = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  pord = 2,
  optim_method = c("ucminf", "grid_search"),
  optim_criterion = c("aic", "bic"),
  lrho = c(0, 0),
  Wprior = NULL,
  ridge = 0,
  control_algorithm = list(),
  par_gridsearch = list()
)

Arguments

data2ts

(optional) an object of class created by the function prepare_data(). Proving this input is the easiest way to use the function fit2ts. However, the user can also provide the input data together with a list of bins, as explained by the following parameters' descriptions.

Y

A matrix (or 3d-array) of event counts of dimension nu by ns (or nu by ns by n).

R

A matrix (or 3d-array) of exposure times of dimension nu by ns (or nu by ns by n).

Z

(optional) A regression matrix of covariates values of dimensions n by p.

bins

a list with the specification for the bins. This is created by the function prepare_data. If a list prepared externally from such function if provided, it should contain the following elements: * bins_u A vector of bins extremes for the time scale u. * midu A vector with the midpoints of the bins over u. * nu The number of bins over u. * bins_s A vector of bins extremes for the time scale s. * mids A vector with the midpoints of the bins over s. * ns The number of bins over s.

Bbases_spec

A list with the specification for the B-splines basis with the following elements:

bdeg The degree of the B-splines basis. Default is 3 (for cubic B-splines).

nseg_u The number of segments for the B-splines over u. Default is 10.

min_u (optional) The lower limit of the domain of Bu. Default is min(bins_u).

max_u (optional) The upper limit of the domain of Bu. Default is max(bins_u).

nseg_s The number of segments for the B-splines over s. Default is 10.

min_s (optional) The lower limit of the domain of Bs. Default is min(bins_s).

max_s (optional) The upper limit of the domain of Bs. Default is max(bins_s).

pord

The order of the penalty. Default is 2.

optim_method

The method to be used for optimization: "ucminf" (default) for the numerical optimization of the AIC (or BIC) and "grid_search" for a grid search of the minimum AIC (or BIC) over a grid of log_10(rho_u) and log_10(rho_s) values.

optim_criterion

The criterion to be used for optimization: "aic" (default) or "bic".

lrho

A vector of two elements if optim_method == "ucminf". Default is c(0,0). A list of two vectors of values for log_10(rho_u) and log_10(rho_s) if optim_method == "grid_search". In the latter case, if a list with two vectors is not provided, a default sequence of values is used for both log_10(rho_u) and log_10(rho_s).

Wprior

An optional matrix of a-priori weights.

ridge

A ridge penalty parameter: default is 0.

control_algorithm

A list with optional values for the parameters of the iterative processes:

maxiter The maximum number of iteration for the IWSL algorithm. Default is 20.

conv_crit The convergence criteria, expressed as difference between estimates at iteration i and i+1. Default is 1e-5.

verbose A Boolean. Default is FALSE. If TRUE monitors the iteration process.

monitor_ev A Boolean. Default is FALSE. If TRUE monitors the evaluation of the model over the log_10(rho_s) values.

par_gridsearch

A list of parameters for the grid_search:

plot_aic A Boolean. Default is FALSE. If TRUE, plot the AIC values over the grid of log_10(rho_u) and log_10(rho_s) values.

plot_bic A Boolean. Default is FALSE. If TRUE, plot the BIC values over the grid of log_10(rho_u) and log_10(rho_s) values.

return_aic A Boolean. Default is TRUE. Return the AIC values.

return_bic A Boolean. Default is TRUE. Return the BIC values.

col The color palette to be used for the AIC/BIC plot. Default is grDevices::gray.colors(n=10).

plot_contour A Boolean. Default is TRUE. Adds white contour lines to the AIC/BIC plot.

mark_optimal A Boolean. Default is TRUE. If the plot of the AIC or BIC values is returned, marks the optimal combination of log_10(rho_u) and log_10(rho_s) in the plot.

main_aic The title of the AIC plot. Default is "AIC grid".

main_bic The title of the BIC plot. Default is "BIC grid".

Value

An object of class h2tsfit with the following elements:

optimal_model A list with :

Alpha The matrix of estimated P-splines coefficients of dimension cu by cs.

Cov_alpha The variance-covariance matrix of the Alpha coefficients, of dimension cucs by cucs.

beta The vector of length p of estimated covariates coefficients (if model with covariates).

Cov_beta The variance-covariance matrix of the beta coefficients, of dimension p by p (if model with covariates).

SE_beta The vector of length p of estimated Standard Errors for the beta coefficients (if model with covariates)..

Eta or Eta0 The matrix of values of the (baseline) linear predictor (log-hazard) of dimension nu by ns.

H The hat-matrix.

deviance The deviance.

ed The effective dimension of the model.

aic The value of the AIC.

bic The value of the BIC.

Bbases a list with the B-spline bases Bu and Bs

optimal_logrho A vector with the optimal values of log10(rho_u) and log10(rho_s).

P_optimal The optimal penalty matrix P.

AIC (if par_gridsearch$return_aic == TRUE) The matrix of AIC values.

BIC (if par_gridsearch$return_bic == TRUE) The matrix of BIC values.