Fit a Varying Coefficient Model (VCM) over two time scales

fitvcm() fits a varying coefficient model over two time scales with P-splines. The model is described in Carollo et al. (2025).

Usage

fitvcm(
  data2ts = NULL,
  Y = NULL,
  R = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  pord = 2,
  kappa = 1e-10,
  lsmpar = c(0, 0),
  control_algorithm = list()
)

Arguments

data2ts

(optional) an object of class "data2ts" created by the function prepare_data(). Proving this input is the easiest way to use the function fitvcm(). However, the user can also provide the input data together with a list of bins, as explained by the following parameters' descriptions.

Y

A matrix (or 3d-array) of event counts of dimension nu by ns (or nu by ns by n).

R

A matrix (or 3d-array) of exposure times of dimension nu by ns (or nu by ns by n).

Z

(optional) A regression matrix of covariates values of dimensions n by p.

bins

a list with the specification for the bins. This is created by the function prepare_data(). If a list prepared externally from such function if provided, it should contain the following elements: * bins_u A vector of bins extremes for the time scale u. * midu A vector with the midpoints of the bins over u. * nu The number of bins over u. * bins_s A vector of bins extremes for the time scale s. * mids A vector with the midpoints of the bins over s. * ns The number of bins over s.

Bbases_spec

A list with the specification for the B-splines basis with the following elements:

bdeg The degree of the B-splines basis. Default is 3 (for cubic B-splines).

nseg_s The number of segments for the B-splines over s. Default is 10.

min_s (optional) The lower limit of the domain of Bs. Default is min(bins_s).

max_s (optional) The upper limit of the domain of Bs. Default is max(bins_s).

pord

The order of the penalty. Default is 2.

kappa

A ridge penalty.

lsmpar

The starting values for the two smoothing parameters. A vector of two elements, default is c(0,0).

control_algorithm

A list with optional values for the parameters of the iterative processes:

maxiter The maximum number of iteration for the IWSL algorithm. Default is 20.

conv_crit The convergence criteria, expressed as difference between estimates at iteration i and i+1. Default is 1e-5.

verbose A Boolean. Default is FALSE. If TRUE monitors the iteration process.

monitor_ev A Boolean. Default is FALSE. If TRUE monitors the evaluation of the model over the log_10(rho_s) values.

Value

An object of class haz2tsVCM, that is a list with the following elements:

optimal_model A list with :

Alpha The matrix of estimated P-splines coefficients of dimension \(c_s\) by 2.

Cov_alpha The variance-covariance matrix of the Alpha coefficients, of dimension \(c_uc_s\) by \(c_uc_s\).

Eta The matrix of values of the baseline linear predictor (log-hazard) of dimension \(n_u\) by \(n_s\).

H The hat-matrix.

deviance The deviance.

ed The effective dimension of the model.

aic The value of the AIC.

bic The value of the BIC.

Bbases a list with the B-spline bases Bu and Bs

optimal_logsmpar A vector with the optimal values of \(\log_{10}(\theta)\) and \(\log_{10}(\phi)\).

P_optimal The optimal penalty matrix P.

AIC (if par_gridsearch$return_aic == TRUE) The matrix of AIC values.

BIC (if par_gridsearch$return_bic == TRUE) The matrix of BIC values.

@references Carollo, A., Putter, H., Eilers, P. H. C., & Gampe, J. (2025). Analysis of Time-to-Event Data With Two Time Scales. An Application to Transitions out of Cohabitation. Sociological Methods & Research, 0(0). https://doi.org/10.1177/00491241251374193

Examples

# Create some fake data - the bare minimum
id <- 1:20
u <- c(
  5.43, 3.25, 8.15, 5.53, 7.28, 6.61, 5.91, 4.94, 4.25, 3.86, 4.05, 6.86,
  4.94, 4.46, 2.14, 7.56, 5.55, 7.60, 6.46, 4.96
)
s <- c(
  0.44, 4.89, 0.92, 1.81, 2.02, 1.55, 3.16, 6.36, 0.66, 2.02, 1.22, 3.96,
  7.07, 2.91, 3.38, 2.36, 1.74, 0.06, 5.76, 3.00
)
ev <- c(1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1) #'

fakedata <- as.data.frame(cbind(id, u, s, ev))
fakedata2ts <- prepare_data(
  data = fakedata,
  u = "u",
  s_out = "s",
  ev = "ev",
  ds = .5
)
#> `s_in = NULL`. I will use `s_in = 0` for all observations.
#> `s_in = NULL`. I will use `s_in = 0` for all observations.
# Fit a fake model - not optimal smoothing
fitvcm(fakedata2ts)
#> $optimal_model
#> $optimal_model$Alpha
#>             [,1]        [,2]
#>  [1,] -1.1942213 -0.25486340
#>  [2,] -1.2292067 -0.21957418
#>  [3,] -1.2530351 -0.17903689
#>  [4,] -1.1597732 -0.08897924
#>  [5,] -1.1829444 -0.04690506
#>  [6,] -1.3297915 -0.08379147
#>  [7,] -2.0222779 -0.22894554
#>  [8,] -2.1419501 -0.31159060
#>  [9,] -1.4357834 -0.28675769
#> [10,] -0.2490027 -0.19569126
#> [11,]  0.4179042 -0.18177843
#> [12,]  0.7087067 -0.22356551
#> [13,]  0.9880183 -0.26705197
#> 
#> $optimal_model$SE_Alpha
#>            [,1]      [,2]
#>  [1,]  5.321617 1.0068583
#>  [2,]  3.062662 0.5916669
#>  [3,]  1.931576 0.3685710
#>  [4,]  1.633242 0.3072392
#>  [5,]  1.799442 0.3465888
#>  [6,]  2.357016 0.4714154
#>  [7,]  3.427269 0.6848802
#>  [8,]  4.931340 0.9744117
#>  [9,]  6.816683 1.3390351
#> [10,]  9.002245 1.7998237
#> [11,] 11.517053 2.3481291
#> [12,] 14.466214 2.9708164
#> [13,] 17.911291 3.6720958
#> 
#> $optimal_model$VarCov
#>              [,1]       [,2]        [,3]        [,4]       [,5]        [,6]
#>  [1,] 28.31960991 14.2013048  4.86864783  0.30426805 -2.0029083 -3.20491148
#>  [2,] 14.20130480  9.3798960  4.32933594  1.12805210 -0.7298012 -1.83965047
#>  [3,]  4.86864783  4.3293359  3.73098412  1.95310525  0.5491054 -0.46235892
#>  [4,]  0.30426805  1.1280521  1.95310525  2.66747879  1.9040420  1.08955740
#>  [5,] -2.00290827 -0.7298012  0.54910544  1.90404204  3.2379915  3.09488794
#>  [6,] -3.20491148 -1.8396505 -0.46235892  1.08955740  3.0948879  5.55552474
#>  [7,] -3.95200402 -2.6249468 -1.27769107  0.36865229  2.8421968  6.72781206
#>  [8,] -4.52256178 -3.2787530 -2.00725440 -0.31292670  2.6030216  7.66696578
#>  [9,] -5.01908373 -3.8752797 -2.69720557 -0.98190769  2.3545602  8.57526733
#> [10,] -5.42656548 -4.4017173 -3.33705622 -1.63624709  2.0559715  9.38043559
#> [11,] -5.78962694 -4.8928556 -3.95122845 -2.28309621  1.7242788 10.11904775
#> [12,] -6.15729598 -5.3875861 -4.56797187 -2.93169905  1.3908003 10.86462982
#> [13,] -6.52561750 -5.8828312 -5.18508761 -3.58048085  1.0574542 11.61121058
#> [14,] -3.76620674 -2.4153493 -1.09331579 -0.10701557  0.3971300  0.62669359
#> [15,] -2.40354321 -1.6283327 -0.86456247 -0.24250686  0.1417167  0.35982756
#> [16,] -1.07174326 -0.8541122 -0.63232180 -0.37333910 -0.1137156  0.09053826
#> [17,] -0.09308935 -0.2339554 -0.37012689 -0.43709461 -0.3623562 -0.21498435
#> [18,]  0.40857264  0.1471239 -0.11475918 -0.37148983 -0.5519932 -0.59999951
#> [19,]  0.65574868  0.3750540  0.09115003 -0.23447655 -0.6216930 -0.99732546
#> [20,]  0.80345986  0.5327709  0.25740179 -0.08714238 -0.6021067 -1.30449194
#> [21,]  0.90796694  0.6571224  0.40049390  0.05452286 -0.5452463 -1.52707007
#> [22,]  0.99383605  0.7667992  0.53294695  0.19152723 -0.4759952 -1.70296444
#> [23,]  1.08455298  0.8802259  0.66795819  0.32856988 -0.4105385 -1.87961697
#> [24,]  1.18133685  0.9984712  0.80646913  0.46658322 -0.3497724 -2.06523217
#> [25,]  1.27743730  1.1161836  0.94459868  0.60433642 -0.2892713 -2.24981354
#> [26,]  1.37344096  1.2338197  1.08267301  0.74206308 -0.2287504 -2.43424677
#>             [,7]         [,8]        [,9]       [,10]       [,11]       [,12]
#>  [1,] -3.9520040  -4.52256178  -5.0190837  -5.4265655  -5.7896269  -6.1572960
#>  [2,] -2.6249468  -3.27875296  -3.8752797  -4.4017173  -4.8928556  -5.3875861
#>  [3,] -1.2776911  -2.00725440  -2.6972056  -3.3370562  -3.9512284  -4.5679719
#>  [4,]  0.3686523  -0.31292670  -0.9819077  -1.6362471  -2.2830962  -2.9316990
#>  [5,]  2.8421968   2.60302160   2.3545602   2.0559715   1.7242788   1.3908003
#>  [6,]  6.7278121   7.66696578   8.5752673   9.3804356  10.1190478  10.8646298
#>  [7,] 11.7461704  15.33300859  18.5637143  21.6186629  24.6376075  27.7050246
#>  [8,] 15.3330086  24.31811806  31.7898583  38.6773357  45.5393540  52.5222216
#>  [9,] 18.5637143  31.78985831  46.4671609  59.3327359  71.8714648  84.5748427
#> [10,] 21.6186629  38.67733567  59.3327359  81.0404068 101.3599923 121.7214399
#> [11,] 24.6376075  45.53935396  71.8714648 101.3599923 132.6425017 163.4853309
#> [12,] 27.7050246  52.52222160  84.5748427 121.7214399 163.4853309 209.2713463
#> [13,] 30.7760171  59.51337416  97.2914756 142.0955398 194.3223856 255.0125788
#> [14,]  0.7680054   0.88051559   0.9809599   1.0641498   1.1381337   1.2128363
#> [15,]  0.5113482   0.63830867   0.7550054   0.8583924   0.9549104   1.0521029
#> [16,]  0.2510735   0.39154891   0.5235777   0.6462737   0.7644581   0.8832635
#> [17,] -0.0695603   0.06630317   0.1967177   0.3225876   0.4465933   0.5709593
#> [18,] -0.5764202  -0.52795355  -0.4720608  -0.4096969  -0.3446527  -0.2802573
#> [19,] -1.3025471  -1.54522779  -1.7505060  -1.9168129  -2.0650909  -2.2148313
#> [20,] -2.1101507  -2.91807770  -3.6833012  -4.3556891  -4.9730339  -5.5890856
#> [21,] -2.8863435  -4.46988404  -6.1111343  -7.6342132  -9.0574149 -10.4731052
#> [22,] -3.6077091  -6.06765349  -8.8336013 -11.5637319 -14.1882496 -16.8040218
#> [23,] -4.3239750  -7.69966829 -11.7471789 -16.0006539 -20.2524154 -24.5190571
#> [24,] -5.0436007  -9.33209748 -14.7042520 -20.6365875 -26.8124115 -33.0799554
#> [25,] -5.7560366 -10.94660133 -17.6369031 -25.2663118 -33.4376245 -41.8151962
#> [26,] -6.4679422 -12.55987629 -20.5675881 -29.8941593 -40.0636939 -50.5570796
#>             [,13]       [,14]       [,15]       [,16]       [,17]       [,18]
#>  [1,]  -6.5256175 -3.76620674 -2.40354321 -1.07174326 -0.09308935  0.40857264
#>  [2,]  -5.8828312 -2.41534932 -1.62833269 -0.85411221 -0.23395536  0.14712388
#>  [3,]  -5.1850876 -1.09331579 -0.86456247 -0.63232180 -0.37012689 -0.11475918
#>  [4,]  -3.5804809 -0.10701557 -0.24250686 -0.37333910 -0.43709461 -0.37148983
#>  [5,]   1.0574542  0.39713002  0.14171666 -0.11371559 -0.36235625 -0.55199321
#>  [6,]  11.6112106  0.62669359  0.35982756  0.09053826 -0.21498435 -0.59999951
#>  [7,]  30.7760171  0.76800539  0.51134821  0.25107346 -0.06956030 -0.57642017
#>  [8,]  59.5133742  0.88051559  0.63830867  0.39154891  0.06630317 -0.52795355
#>  [9,]  97.2914756  0.98095985  0.75500538  0.52357767  0.19671767 -0.47206081
#> [10,] 142.0955398  1.06414977  0.85839243  0.64627374  0.32258763 -0.40969686
#> [11,] 194.3223856  1.13813373  0.95491045  0.76445815  0.44659330 -0.34465275
#> [12,] 255.0125788  1.21283628  1.05210292  0.88326350  0.57095925 -0.28025730
#> [13,] 320.8143603  1.28765834  1.14939399  1.00214572  0.69536392 -0.21592612
#> [14,]   1.2876583  1.01376363  0.52931410  0.19155598  0.01490649 -0.07559903
#> [15,]   1.1493940  0.52931410  0.35006973  0.16460023  0.04327139 -0.02777619
#> [16,]   1.0021457  0.19155598  0.16460023  0.13584456  0.07151421  0.02016607
#> [17,]   0.6953639  0.01490649  0.04327139  0.07151421  0.09439591  0.07003633
#> [18,]  -0.2159261 -0.07559903 -0.02777619  0.02016607  0.07003633  0.12012381
#> [19,]  -2.3648069 -0.12567615 -0.07215786 -0.01827199  0.04191373  0.12279227
#> [20,]  -6.2054708 -0.15624758 -0.10337443 -0.04986366  0.01444404  0.11552833
#> [21,] -11.8890706 -0.17748021 -0.12792343 -0.07749431 -0.01214773  0.10515047
#> [22,] -19.4200787 -0.19467666 -0.14947264 -0.10319204 -0.03826007  0.09326174
#> [23,] -28.7868552 -0.21278608 -0.17169840 -0.12933456 -0.06462803  0.08164628
#> [24,] -39.3521732 -0.23215826 -0.19486154 -0.15608665 -0.09124505  0.07038656
#> [25,] -50.2040122 -0.25142386 -0.21792464 -0.18274663 -0.11780864  0.05903062
#> [26,] -61.0685959 -0.27067174 -0.24097311 -0.20939521 -0.14436648  0.04766515
#>             [,19]       [,20]        [,21]        [,22]        [,23]
#>  [1,]  0.65574868  0.80345986   0.90796694   0.99383605   1.08455298
#>  [2,]  0.37505400  0.53277086   0.65712242   0.76679916   0.88022591
#>  [3,]  0.09115003  0.25740179   0.40049390   0.53294695   0.66795819
#>  [4,] -0.23447655 -0.08714238   0.05452286   0.19152723   0.32856988
#>  [5,] -0.62169300 -0.60210671  -0.54524633  -0.47599516  -0.41053851
#>  [6,] -0.99732546 -1.30449194  -1.52707007  -1.70296444  -1.87961697
#>  [7,] -1.30254714 -2.11015075  -2.88634350  -3.60770909  -4.32397497
#>  [8,] -1.54522779 -2.91807770  -4.46988404  -6.06765349  -7.69966829
#>  [9,] -1.75050598 -3.68330121  -6.11113430  -8.83360132 -11.74717891
#> [10,] -1.91681287 -4.35568914  -7.63421324 -11.56373193 -16.00065395
#> [11,] -2.06509093 -4.97303387  -9.05741490 -14.18824957 -20.25241540
#> [12,] -2.21483134 -5.58908559 -10.47310524 -16.80402176 -24.51905708
#> [13,] -2.36480686 -6.20547082 -11.88907057 -19.42007868 -28.78685518
#> [14,] -0.12567615 -0.15624758  -0.17748021  -0.19467666  -0.21278608
#> [15,] -0.07215786 -0.10337443  -0.12792343  -0.14947264  -0.17169840
#> [16,] -0.01827199 -0.04986366  -0.07749431  -0.10319204  -0.12933456
#> [17,]  0.04191373  0.01444404  -0.01214773  -0.03826007  -0.06462803
#> [18,]  0.12279227  0.11552833   0.10515047   0.09326174   0.08164628
#> [19,]  0.22223253  0.27508495   0.31315300   0.34643291   0.38256644
#> [20,]  0.27508495  0.46906090   0.61184863   0.73808162   0.87143790
#> [21,]  0.31315300  0.61184863   0.94947809   1.23881311   1.53392574
#> [22,]  0.34643291  0.73808162   1.23881311   1.79301495   2.33484866
#> [23,]  0.38256644  0.87143790   1.53392574   2.33484866   3.23936543
#> [24,]  0.42119137  1.01254191   1.84311331   2.89122531   4.14262879
#> [25,]  0.45959940  1.15383771   2.15341500   3.44889915   5.04368500
#> [26,]  0.49797255  1.29508403   2.46367591   4.00653075   5.94456967
#>              [,24]        [,25]        [,26]
#>  [1,]   1.18133685   1.27743730   1.37344096
#>  [2,]   0.99847118   1.11618361   1.23381973
#>  [3,]   0.80646913   0.94459868   1.08267301
#>  [4,]   0.46658322   0.60433642   0.74206308
#>  [5,]  -0.34977244  -0.28927127  -0.22875045
#>  [6,]  -2.06523217  -2.24981354  -2.43424677
#>  [7,]  -5.04360072  -5.75603663  -6.46794219
#>  [8,]  -9.33209748 -10.94660133 -12.55987629
#>  [9,] -14.70425204 -17.63690313 -20.56758811
#> [10,] -20.63658753 -25.26631178 -29.89415933
#> [11,] -26.81241153 -33.43762446 -40.06369388
#> [12,] -33.07995542 -41.81519624 -50.55707959
#> [13,] -39.35217319 -50.20401224 -61.06859592
#> [14,]  -0.23215826  -0.25142386  -0.27067174
#> [15,]  -0.19486154  -0.21792464  -0.24097311
#> [16,]  -0.15608665  -0.18274663  -0.20939521
#> [17,]  -0.09124505  -0.11780864  -0.14436648
#> [18,]   0.07038656   0.05903062   0.04766515
#> [19,]   0.42119137   0.45959940   0.49797255
#> [20,]   1.01254191   1.15383771   1.29508403
#> [21,]   1.84311331   2.15341500   2.46367591
#> [22,]   2.89122531   3.44889915   4.00653075
#> [23,]   4.14262879   5.04368500   5.94456967
#> [24,]   5.51371013   6.87118685   8.22797029
#> [25,]   6.87118685   8.82575027  10.77864586
#> [26,]   8.22797029  10.77864586  13.48428770
#> 
#> $optimal_model$deviance
#> [1] 31.53268
#> 
#> $optimal_model$Eta
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] -1.727013 -1.641693 -1.485226 -1.359980 -1.347113 -1.493370 -1.911787
#>  [2,] -1.829762 -1.727084 -1.545758 -1.398885 -1.377145 -1.532035 -1.980355
#>  [3,] -1.932510 -1.812476 -1.606289 -1.437791 -1.407178 -1.570700 -2.048923
#>  [4,] -2.035259 -1.897868 -1.666820 -1.476697 -1.437210 -1.609365 -2.117492
#>  [5,] -2.138007 -1.983260 -1.727351 -1.515603 -1.467243 -1.648030 -2.186060
#>  [6,] -2.240756 -2.068651 -1.787882 -1.554508 -1.497276 -1.686695 -2.254629
#>  [7,] -2.343504 -2.154043 -1.848413 -1.593414 -1.527308 -1.725360 -2.323197
#>  [8,] -2.446253 -2.239435 -1.908945 -1.632320 -1.557341 -1.764024 -2.391766
#>  [9,] -2.549001 -2.324827 -1.969476 -1.671225 -1.587373 -1.802689 -2.460334
#> [10,] -2.651749 -2.410218 -2.030007 -1.710131 -1.617406 -1.841354 -2.528902
#> [11,] -2.754498 -2.495610 -2.090538 -1.749037 -1.647439 -1.880019 -2.597471
#> [12,] -2.857246 -2.581002 -2.151069 -1.787943 -1.677471 -1.918684 -2.666039
#> [13,] -2.959995 -2.666394 -2.211600 -1.826848 -1.707504 -1.957349 -2.734608
#>            [,8]      [,9]     [,10]     [,11]     [,12]      [,13]      [,14]
#>  [1,] -2.449089 -2.722995 -2.574068 -2.014649 -1.215683 -0.5212408 -0.1014175
#>  [2,] -2.558353 -2.862169 -2.722866 -2.152508 -1.330407 -0.6188446 -0.1969483
#>  [3,] -2.667617 -3.001344 -2.871665 -2.290368 -1.445131 -0.7164483 -0.2924792
#>  [4,] -2.776880 -3.140518 -3.020463 -2.428227 -1.559855 -0.8140521 -0.3880101
#>  [5,] -2.886144 -3.279693 -3.169261 -2.566086 -1.674579 -0.9116559 -0.4835410
#>  [6,] -2.995408 -3.418867 -3.318059 -2.703946 -1.789303 -1.0092596 -0.5790718
#>  [7,] -3.104671 -3.558042 -3.466858 -2.841805 -1.904027 -1.1068634 -0.6746027
#>  [8,] -3.213935 -3.697216 -3.615656 -2.979665 -2.018751 -1.2044672 -0.7701336
#>  [9,] -3.323199 -3.836390 -3.764454 -3.117524 -2.133475 -1.3020709 -0.8656644
#> [10,] -3.432462 -3.975565 -3.913253 -3.255383 -2.248199 -1.3996747 -0.9611953
#> [11,] -3.541726 -4.114739 -4.062051 -3.393243 -2.362923 -1.4972785 -1.0567262
#> [12,] -3.650990 -4.253914 -4.210849 -3.531102 -2.477647 -1.5948822 -1.1522571
#> [13,] -3.760253 -4.393088 -4.359647 -3.668962 -2.592371 -1.6924860 -1.2477879
#>              [,15]
#>  [1,]  0.106829949
#>  [2,]  0.001797833
#>  [3,] -0.103234282
#>  [4,] -0.208266398
#>  [5,] -0.313298514
#>  [6,] -0.418330630
#>  [7,] -0.523362746
#>  [8,] -0.628394862
#>  [9,] -0.733426978
#> [10,] -0.838459094
#> [11,] -0.943491209
#> [12,] -1.048523325
#> [13,] -1.153555441
#> 
#> $optimal_model$ed
#> [1] 7.853223
#> 
#> $optimal_model$aic
#> [1] 47.23913
#> 
#> $optimal_model$bic
#> [1] 66.04327
#> 
#> $optimal_model$Bbases
#> $optimal_model$Bbases$bdeg
#> [1] 3
#> 
#> $optimal_model$Bbases$nseg_s
#> [1] 10
#> 
#> $optimal_model$Bbases$min_s
#> [1] 0
#> 
#> $optimal_model$Bbases$max_s
#> [1] 7.5
#> 
#> $optimal_model$Bbases$Bs
#>              [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>   [1,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [2,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [3,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [4,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [5,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [6,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [7,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [8,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>   [9,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>  [10,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>  [11,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>  [12,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>  [13,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#>  [14,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [15,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [16,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [17,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [18,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [19,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [20,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [21,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [22,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [23,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [24,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [25,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [26,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#>  [27,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [28,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [29,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [30,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [31,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [32,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [33,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [34,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [35,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [36,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [37,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [38,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [39,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#>  [40,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [41,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [42,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [43,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [44,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [45,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [46,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [47,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [48,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [49,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [50,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [51,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [52,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#>  [53,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [54,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [55,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [56,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [57,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [58,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [59,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [60,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [61,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [62,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [63,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [64,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [65,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#>  [66,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [67,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [68,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [69,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [70,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [71,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [72,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [73,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [74,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [75,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [76,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [77,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [78,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>  [79,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [80,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [81,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [82,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [83,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [84,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [85,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [86,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [87,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [88,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [89,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [90,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [91,] 0.00000000 0.00000000 0.00000000 0.00000000 0.04938272 0.57407407
#>  [92,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [93,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [94,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [95,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [96,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [97,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [98,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#>  [99,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [100,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [101,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [102,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [103,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [104,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667
#> [105,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [106,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [107,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [108,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [109,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [110,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [111,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [112,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [113,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [114,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [115,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [116,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [117,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00617284
#> [118,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [119,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [120,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [121,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [122,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [123,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [124,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [125,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [126,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [127,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [128,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [129,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [130,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [131,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [132,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [133,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [134,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [135,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [136,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [137,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [138,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [139,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [140,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [141,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [142,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [143,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [144,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [145,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [146,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [147,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [148,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [149,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [150,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [151,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [152,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [153,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [154,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [155,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [156,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [157,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [158,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [159,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [160,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [161,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [162,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [163,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [164,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [165,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [166,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [167,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [168,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [169,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [170,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [171,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [172,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [173,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [174,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [175,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [176,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [177,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [178,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [179,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [180,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [181,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [182,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [183,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [184,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [185,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [186,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [187,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [188,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [189,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [190,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [191,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [192,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [193,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [194,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [195,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>              [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>   [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [5,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [6,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [10,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [11,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [12,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [14,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [20,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [21,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [22,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [23,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [26,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [27,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [28,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [29,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [30,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [32,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [33,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [34,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [35,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [36,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [38,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [39,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [40,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [41,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [42,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [44,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [45,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [46,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [47,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [48,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [50,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [51,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [52,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [53,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [54,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [56,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [57,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [58,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [59,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [60,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [61,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [62,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [63,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [64,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [65,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [66,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [68,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [69,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [70,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [71,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [72,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [74,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [75,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [76,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [77,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [78,] 0.04938272 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [79,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [80,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [81,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [82,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [83,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [84,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [85,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [86,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [87,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [88,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [89,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [90,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [91,] 0.37037037 0.00617284 0.00000000 0.00000000 0.00000000 0.00000000
#>  [92,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [93,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [94,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [95,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [96,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [97,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [98,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#>  [99,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [100,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [101,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [102,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [103,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [104,] 0.66666667 0.16666667 0.00000000 0.00000000 0.00000000 0.00000000
#> [105,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [106,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [107,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [108,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [109,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [110,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [111,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [112,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [113,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [114,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [115,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [116,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [117,] 0.37037037 0.57407407 0.04938272 0.00000000 0.00000000 0.00000000
#> [118,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [119,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [120,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [121,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [122,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [123,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [124,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [125,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [126,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [127,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [128,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [129,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [130,] 0.04938272 0.57407407 0.37037037 0.00617284 0.00000000 0.00000000
#> [131,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [132,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [133,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [134,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [135,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [136,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [137,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [138,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [139,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [140,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [141,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [142,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [143,] 0.00000000 0.16666667 0.66666667 0.16666667 0.00000000 0.00000000
#> [144,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [145,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [146,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [147,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [148,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [149,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [150,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [151,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [152,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [153,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [154,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [155,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [156,] 0.00000000 0.00617284 0.37037037 0.57407407 0.04938272 0.00000000
#> [157,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [158,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [159,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [160,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [161,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [162,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [163,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [164,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [165,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [166,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [167,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [168,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [169,] 0.00000000 0.00000000 0.04938272 0.57407407 0.37037037 0.00617284
#> [170,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [171,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [172,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [173,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [174,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [175,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [176,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [177,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [178,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [179,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [180,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [181,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [182,] 0.00000000 0.00000000 0.00000000 0.16666667 0.66666667 0.16666667
#> [183,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [184,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [185,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [186,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [187,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [188,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [189,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [190,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [191,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [192,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [193,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [194,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#> [195,] 0.00000000 0.00000000 0.00000000 0.00617284 0.37037037 0.57407407
#>             [,13]
#>   [1,] 0.00000000
#>   [2,] 0.00000000
#>   [3,] 0.00000000
#>   [4,] 0.00000000
#>   [5,] 0.00000000
#>   [6,] 0.00000000
#>   [7,] 0.00000000
#>   [8,] 0.00000000
#>   [9,] 0.00000000
#>  [10,] 0.00000000
#>  [11,] 0.00000000
#>  [12,] 0.00000000
#>  [13,] 0.00000000
#>  [14,] 0.00000000
#>  [15,] 0.00000000
#>  [16,] 0.00000000
#>  [17,] 0.00000000
#>  [18,] 0.00000000
#>  [19,] 0.00000000
#>  [20,] 0.00000000
#>  [21,] 0.00000000
#>  [22,] 0.00000000
#>  [23,] 0.00000000
#>  [24,] 0.00000000
#>  [25,] 0.00000000
#>  [26,] 0.00000000
#>  [27,] 0.00000000
#>  [28,] 0.00000000
#>  [29,] 0.00000000
#>  [30,] 0.00000000
#>  [31,] 0.00000000
#>  [32,] 0.00000000
#>  [33,] 0.00000000
#>  [34,] 0.00000000
#>  [35,] 0.00000000
#>  [36,] 0.00000000
#>  [37,] 0.00000000
#>  [38,] 0.00000000
#>  [39,] 0.00000000
#>  [40,] 0.00000000
#>  [41,] 0.00000000
#>  [42,] 0.00000000
#>  [43,] 0.00000000
#>  [44,] 0.00000000
#>  [45,] 0.00000000
#>  [46,] 0.00000000
#>  [47,] 0.00000000
#>  [48,] 0.00000000
#>  [49,] 0.00000000
#>  [50,] 0.00000000
#>  [51,] 0.00000000
#>  [52,] 0.00000000
#>  [53,] 0.00000000
#>  [54,] 0.00000000
#>  [55,] 0.00000000
#>  [56,] 0.00000000
#>  [57,] 0.00000000
#>  [58,] 0.00000000
#>  [59,] 0.00000000
#>  [60,] 0.00000000
#>  [61,] 0.00000000
#>  [62,] 0.00000000
#>  [63,] 0.00000000
#>  [64,] 0.00000000
#>  [65,] 0.00000000
#>  [66,] 0.00000000
#>  [67,] 0.00000000
#>  [68,] 0.00000000
#>  [69,] 0.00000000
#>  [70,] 0.00000000
#>  [71,] 0.00000000
#>  [72,] 0.00000000
#>  [73,] 0.00000000
#>  [74,] 0.00000000
#>  [75,] 0.00000000
#>  [76,] 0.00000000
#>  [77,] 0.00000000
#>  [78,] 0.00000000
#>  [79,] 0.00000000
#>  [80,] 0.00000000
#>  [81,] 0.00000000
#>  [82,] 0.00000000
#>  [83,] 0.00000000
#>  [84,] 0.00000000
#>  [85,] 0.00000000
#>  [86,] 0.00000000
#>  [87,] 0.00000000
#>  [88,] 0.00000000
#>  [89,] 0.00000000
#>  [90,] 0.00000000
#>  [91,] 0.00000000
#>  [92,] 0.00000000
#>  [93,] 0.00000000
#>  [94,] 0.00000000
#>  [95,] 0.00000000
#>  [96,] 0.00000000
#>  [97,] 0.00000000
#>  [98,] 0.00000000
#>  [99,] 0.00000000
#> [100,] 0.00000000
#> [101,] 0.00000000
#> [102,] 0.00000000
#> [103,] 0.00000000
#> [104,] 0.00000000
#> [105,] 0.00000000
#> [106,] 0.00000000
#> [107,] 0.00000000
#> [108,] 0.00000000
#> [109,] 0.00000000
#> [110,] 0.00000000
#> [111,] 0.00000000
#> [112,] 0.00000000
#> [113,] 0.00000000
#> [114,] 0.00000000
#> [115,] 0.00000000
#> [116,] 0.00000000
#> [117,] 0.00000000
#> [118,] 0.00000000
#> [119,] 0.00000000
#> [120,] 0.00000000
#> [121,] 0.00000000
#> [122,] 0.00000000
#> [123,] 0.00000000
#> [124,] 0.00000000
#> [125,] 0.00000000
#> [126,] 0.00000000
#> [127,] 0.00000000
#> [128,] 0.00000000
#> [129,] 0.00000000
#> [130,] 0.00000000
#> [131,] 0.00000000
#> [132,] 0.00000000
#> [133,] 0.00000000
#> [134,] 0.00000000
#> [135,] 0.00000000
#> [136,] 0.00000000
#> [137,] 0.00000000
#> [138,] 0.00000000
#> [139,] 0.00000000
#> [140,] 0.00000000
#> [141,] 0.00000000
#> [142,] 0.00000000
#> [143,] 0.00000000
#> [144,] 0.00000000
#> [145,] 0.00000000
#> [146,] 0.00000000
#> [147,] 0.00000000
#> [148,] 0.00000000
#> [149,] 0.00000000
#> [150,] 0.00000000
#> [151,] 0.00000000
#> [152,] 0.00000000
#> [153,] 0.00000000
#> [154,] 0.00000000
#> [155,] 0.00000000
#> [156,] 0.00000000
#> [157,] 0.00000000
#> [158,] 0.00000000
#> [159,] 0.00000000
#> [160,] 0.00000000
#> [161,] 0.00000000
#> [162,] 0.00000000
#> [163,] 0.00000000
#> [164,] 0.00000000
#> [165,] 0.00000000
#> [166,] 0.00000000
#> [167,] 0.00000000
#> [168,] 0.00000000
#> [169,] 0.00000000
#> [170,] 0.00000000
#> [171,] 0.00000000
#> [172,] 0.00000000
#> [173,] 0.00000000
#> [174,] 0.00000000
#> [175,] 0.00000000
#> [176,] 0.00000000
#> [177,] 0.00000000
#> [178,] 0.00000000
#> [179,] 0.00000000
#> [180,] 0.00000000
#> [181,] 0.00000000
#> [182,] 0.00000000
#> [183,] 0.04938272
#> [184,] 0.04938272
#> [185,] 0.04938272
#> [186,] 0.04938272
#> [187,] 0.04938272
#> [188,] 0.04938272
#> [189,] 0.04938272
#> [190,] 0.04938272
#> [191,] 0.04938272
#> [192,] 0.04938272
#> [193,] 0.04938272
#> [194,] 0.04938272
#> [195,] 0.04938272
#> attr(,"x")
#>   [1] 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.75 0.75
#>  [16] 0.75 0.75 0.75 0.75 0.75 0.75 0.75 0.75 0.75 0.75 0.75 1.25 1.25 1.25 1.25
#>  [31] 1.25 1.25 1.25 1.25 1.25 1.25 1.25 1.25 1.25 1.75 1.75 1.75 1.75 1.75 1.75
#>  [46] 1.75 1.75 1.75 1.75 1.75 1.75 1.75 2.25 2.25 2.25 2.25 2.25 2.25 2.25 2.25
#>  [61] 2.25 2.25 2.25 2.25 2.25 2.75 2.75 2.75 2.75 2.75 2.75 2.75 2.75 2.75 2.75
#>  [76] 2.75 2.75 2.75 3.25 3.25 3.25 3.25 3.25 3.25 3.25 3.25 3.25 3.25 3.25 3.25
#>  [91] 3.25 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 3.75 4.25
#> [106] 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.25 4.75 4.75 4.75
#> [121] 4.75 4.75 4.75 4.75 4.75 4.75 4.75 4.75 4.75 4.75 5.25 5.25 5.25 5.25 5.25
#> [136] 5.25 5.25 5.25 5.25 5.25 5.25 5.25 5.25 5.75 5.75 5.75 5.75 5.75 5.75 5.75
#> [151] 5.75 5.75 5.75 5.75 5.75 5.75 6.25 6.25 6.25 6.25 6.25 6.25 6.25 6.25 6.25
#> [166] 6.25 6.25 6.25 6.25 6.75 6.75 6.75 6.75 6.75 6.75 6.75 6.75 6.75 6.75 6.75
#> [181] 6.75 6.75 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25 7.25
#> attr(,"xl")
#> [1] 0
#> attr(,"xr")
#> [1] 7.5
#> attr(,"nseg")
#> [1] 10
#> attr(,"bdeg")
#> [1] 3
#> attr(,"type")
#> [1] "bbase"
#> 
#> 
#> 
#> $optimal_logsmpar
#> [1] -0.7157954  0.8022907
#> 
#> $P_optim
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    1   -2    1    0    0    0    0    0    0     0     0     0     0
#>  [2,]   -2    5   -4    1    0    0    0    0    0     0     0     0     0
#>  [3,]    1   -4    6   -4    1    0    0    0    0     0     0     0     0
#>  [4,]    0    1   -4    6   -4    1    0    0    0     0     0     0     0
#>  [5,]    0    0    1   -4    6   -4    1    0    0     0     0     0     0
#>  [6,]    0    0    0    1   -4    6   -4    1    0     0     0     0     0
#>  [7,]    0    0    0    0    1   -4    6   -4    1     0     0     0     0
#>  [8,]    0    0    0    0    0    1   -4    6   -4     1     0     0     0
#>  [9,]    0    0    0    0    0    0    1   -4    6    -4     1     0     0
#> [10,]    0    0    0    0    0    0    0    1   -4     6    -4     1     0
#> [11,]    0    0    0    0    0    0    0    0    1    -4     6    -4     1
#> [12,]    0    0    0    0    0    0    0    0    0     1    -4     5    -2
#> [13,]    0    0    0    0    0    0    0    0    0     0     1    -2     1
#> 
#> $nevents
#> [1] 8
#> 
#> $ns
#> [1] 13
#> 
#> $cs
#> [1] 13
#> 
#> $ufitting
#>  [1] 2.39 2.89 3.39 3.89 4.39 4.89 5.39 5.89 6.39 6.89 7.39 7.89 8.39
#> 
#> attr(,"class")
#> [1] "haz2tsVCM"