Get estimated (log-)hazard values with 1 time scale

get_hazard_1d_LMM() takes as input the results of a model estimated by fit1ts() and it returns the estimated values of the smooth log-hazard and the smooth hazard together with their standard errors. This function is specific for objects of class "haz1tsLMM".

If the model includes covariates, then only the baseline (log-)hazard is returned. It is possible to provide values that define a new grid for evaluation of the estimated hazard. If not specified, the hazard is evaluated on the same grid used for the binning of the data, and therefore the estimation of the model. The function will check if the parameters for the new grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Usage

get_hazard_1d_LMM(fitted_model, plot_grid = NULL)

Arguments

fitted_model

is an object of class "haz1tsLMM", the output of the function fit1ts().

plot_grid

(optional) A named vector containing the parameters to build a new grid of intervals over s for plotting the estimated hazard on a finer grid. This must be of the form: plot_grid = c(smin, smax, ds), where smin, smax are the minimum and maximum values desired for the intervals over s, and ds is the distance between intervals over s. If not specified, the plotting is done using the same B-splines basis as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Value

A list with the following elements:

new_plot_grid A list of parameters that specify the new grid, of the form list("ints", "smin", "smax", "ds")

hazard A vector containing the estimated hazard values.

loghazard A vector containing the estimated log-hazard values.

log10hazard A vector containing the estimated log10-hazard values.

SE_hazard A vector containing the estimated SEs for the hazard.

SE_loghazard A vector containing the estimated SEs for the log-hazard.

SE_log10hazard A vector containing the estimated SEs for the log10-hazard.

Examples

## preparing data - no covariates
dt1ts <- prepare_data(
  data = reccolon2ts,
  s_in = "entrys",
  s_out = "timesr",
  events = "status",
  ds = 180
)

## fitting the model with fit1ts()

mod1 <- fit1ts(dt1ts,
  optim_method = "LMMsolver"
)
# Obtain 1d hazard
get_hazard_1d_LMM(mod1)
#> $new_plot_grid
#> $new_plot_grid$ints
#>  [1]    0  288  576  864 1152 1440 1728 2016 2304 2592 2880
#> 
#> $new_plot_grid$smin
#> [1] 0
#> 
#> $new_plot_grid$smax
#> [1] 2880
#> 
#> $new_plot_grid$ds
#> [1] 288
#> 
#> 
#> $hazard
#>  [1] 1.139691e-03 2.126113e-03 1.984001e-03 2.142388e-03 7.897401e-04
#>  [6] 5.380209e-04 7.518546e-04 5.167962e-04 4.019544e-04 1.576618e-04
#> [11] 4.363257e-05
#> 
#> $loghazard
#>  [1]  -6.776998  -6.153460  -6.222640  -6.145834  -7.143807  -7.527613
#>  [7]  -7.192968  -7.567862  -7.819172  -8.755058 -10.039707
#> 
#> $log10hazard
#>  [1] -2.943213 -2.672414 -2.702458 -2.669102 -3.102516 -3.269201 -3.123866
#>  [8] -3.286681 -3.395823 -3.802274 -4.360189
#> 
#> $SE_hazard
#>  [1] 0.0004620736 0.0001936212 0.0002313492 0.0003315360 0.0002227658
#>  [6] 0.0002130237 0.0003189100 0.0003455692 0.0004770558 0.0004738647
#> [11] 0.0003148611
#> 
#> $SE_loghazard
#>  [1] 0.40543750 0.09106816 0.11660742 0.15475070 0.28207478 0.39593948
#>  [7] 0.42416442 0.66867610 1.18684044 3.00557753 7.21619373
#> 
#> $SE_log10hazard
#>  [1] 0.17607927 0.03955040 0.05064196 0.06720737 0.12250352 0.17195433
#>  [7] 0.18421227 0.29040234 0.51543825 1.30530574 3.13395312
#> 
# Change grid
get_hazard_1d_LMM(mod1,
  plot_grid = c(smin = 0, smax = 2730, ds = 30)
)
#> $new_plot_grid
#> $new_plot_grid$ints
#>  [1]    0   30   60   90  120  150  180  210  240  270  300  330  360  390  420
#> [16]  450  480  510  540  570  600  630  660  690  720  750  780  810  840  870
#> [31]  900  930  960  990 1020 1050 1080 1110 1140 1170 1200 1230 1260 1290 1320
#> [46] 1350 1380 1410 1440 1470 1500 1530 1560 1590 1620 1650 1680 1710 1740 1770
#> [61] 1800 1830 1860 1890 1920 1950 1980 2010 2040 2070 2100 2130 2160 2190 2220
#> [76] 2250 2280 2310 2340 2370 2400 2430 2460 2490 2520 2550 2580 2610 2640 2670
#> [91] 2700 2730
#> 
#> $new_plot_grid$smin
#> smin 
#>    0 
#> 
#> $new_plot_grid$smax
#> smax 
#> 2730 
#> 
#> $new_plot_grid$ds
#> ds 
#> 30 
#> 
#> 
#> $hazard
#>  [1] 1.139691e-03 1.307896e-03 1.470508e-03 1.622112e-03 1.758022e-03
#>  [6] 1.874604e-03 1.969461e-03 2.041499e-03 2.090865e-03 2.118795e-03
#> [11] 2.127412e-03 2.119943e-03 2.100592e-03 2.073706e-03 2.043546e-03
#> [16] 2.014173e-03 1.989425e-03 1.972969e-03 1.968421e-03 1.979541e-03
#> [21] 2.009564e-03 2.055347e-03 2.110585e-03 2.168406e-03 2.221178e-03
#> [26] 2.260565e-03 2.277858e-03 2.264632e-03 2.213685e-03 2.120168e-03
#> [31] 1.986720e-03 1.827120e-03 1.655317e-03 1.482854e-03 1.318373e-03
#> [36] 1.167671e-03 1.034104e-03 9.191558e-04 8.230286e-04 7.450803e-04
#> [41] 6.830441e-04 6.341322e-04 5.962209e-04 5.677297e-04 5.475125e-04
#> [46] 5.347800e-04 5.290489e-04 5.301111e-04 5.380209e-04 5.528098e-04
#> [51] 5.735465e-04 5.989937e-04 6.277416e-04 6.580953e-04 6.880031e-04
#> [56] 7.150393e-04 7.364643e-04 7.493765e-04 7.510116e-04 7.408417e-04
#> [61] 7.211909e-04 6.947479e-04 6.641444e-04 6.317746e-04 5.996976e-04
#> [66] 5.696119e-04 5.428859e-04 5.206272e-04 5.035592e-04 4.907897e-04
#> [71] 4.808941e-04 4.726043e-04 4.647560e-04 4.562643e-04 4.461256e-04
#> [76] 4.334430e-04 4.174706e-04 3.976729e-04 3.741451e-04 3.479482e-04
#> [81] 3.202015e-04 2.919040e-04 2.638992e-04 2.368598e-04 2.112873e-04
#> [86] 1.875245e-04 1.657749e-04 1.461222e-04 1.285034e-04 1.127788e-04
#> [91] 9.880157e-05 8.642385e-05
#> 
#> $loghazard
#>  [1] -6.776998 -6.639335 -6.522147 -6.424027 -6.343566 -6.279358 -6.229995
#>  [8] -6.194071 -6.170178 -6.156908 -6.152849 -6.156366 -6.165536 -6.178418
#> [15] -6.193069 -6.207546 -6.219910 -6.228216 -6.230523 -6.224890 -6.209837
#> [22] -6.187311 -6.160790 -6.133763 -6.109718 -6.092141 -6.084520 -6.090343
#> [29] -6.113097 -6.156260 -6.221270 -6.305014 -6.403763 -6.513786 -6.631357
#> [36] -6.752744 -6.874220 -6.992055 -7.102520 -7.202019 -7.288951 -7.363253
#> [43] -7.424899 -7.473865 -7.510125 -7.533655 -7.544430 -7.542424 -7.527613
#> [50] -7.500497 -7.463671 -7.420259 -7.373382 -7.326161 -7.281717 -7.243173
#> [57] -7.213650 -7.196269 -7.194089 -7.207724 -7.234607 -7.271961 -7.317011
#> [64] -7.366978 -7.419085 -7.470555 -7.518611 -7.560476 -7.593809 -7.619495
#> [71] -7.639863 -7.657252 -7.673998 -7.692438 -7.714910 -7.743750 -7.781296
#> [78] -7.829881 -7.890867 -7.963457 -8.046560 -8.139086 -8.239943 -8.348042
#> [85] -8.462292 -8.581601 -8.704880 -8.831067 -8.959555 -9.090083 -9.222397
#> [92] -9.356247
#> 
#> $log10hazard
#>  [1] -2.943213 -2.883427 -2.832533 -2.789919 -2.754976 -2.727090 -2.705653
#>  [8] -2.690051 -2.679674 -2.673911 -2.672148 -2.673676 -2.677658 -2.683253
#> [15] -2.689615 -2.695903 -2.701272 -2.704880 -2.705882 -2.703435 -2.696898
#> [22] -2.687115 -2.675597 -2.663859 -2.653417 -2.645783 -2.642473 -2.645002
#> [29] -2.654884 -2.673630 -2.701863 -2.738233 -2.781119 -2.828902 -2.879962
#> [36] -2.932680 -2.985436 -3.036611 -3.084585 -3.127797 -3.165551 -3.197820
#> [43] -3.224593 -3.245858 -3.261606 -3.271825 -3.276504 -3.275633 -3.269201
#> [50] -3.257424 -3.241431 -3.222578 -3.202219 -3.181711 -3.162410 -3.145670
#> [57] -3.132848 -3.125300 -3.124353 -3.130275 -3.141950 -3.158173 -3.177737
#> [64] -3.199438 -3.222068 -3.244421 -3.265291 -3.283473 -3.297949 -3.309105
#> [71] -3.317951 -3.325502 -3.332775 -3.340784 -3.350543 -3.363068 -3.379374
#> [78] -3.400474 -3.426960 -3.458485 -3.494577 -3.534760 -3.578562 -3.625509
#> [85] -3.675127 -3.726942 -3.780481 -3.835284 -3.891085 -3.947773 -4.005236
#> [92] -4.063366
#> 
#> $SE_hazard
#>  [1] 0.0004620736 0.0003247076 0.0002105086 0.0001521376 0.0001587714
#>  [6] 0.0001837497 0.0001973527 0.0001978318 0.0001937614 0.0001927791
#> [11] 0.0001938614 0.0001914821 0.0001856581 0.0001801539 0.0001792728
#> [16] 0.0001851985 0.0001968111 0.0002107193 0.0002230547 0.0002306085
#> [21] 0.0002318612 0.0002296901 0.0002296345 0.0002372633 0.0002552407
#> [26] 0.0002811551 0.0003084641 0.0003292850 0.0003370171 0.0003285942
#> [31] 0.0003084065 0.0002878006 0.0002744691 0.0002690875 0.0002674166
#> [36] 0.0002645974 0.0002576755 0.0002458624 0.0002299229 0.0002117326
#> [41] 0.0001947169 0.0001821667 0.0001758353 0.0001758267 0.0001808868
#> [46] 0.0001890639 0.0001983153 0.0002068076 0.0002130237 0.0002163023
#> [51] 0.0002181193 0.0002208933 0.0002272564 0.0002391747 0.0002568188
#> [56] 0.0002780002 0.0002986371 0.0003139649 0.0003201634 0.0003180590
#> [61] 0.0003130991 0.0003102200 0.0003119033 0.0003178346 0.0003259663
#> [66] 0.0003340000 0.0003404025 0.0003448470 0.0003484424 0.0003535569
#> [71] 0.0003622921 0.0003756641 0.0003933620 0.0004138656 0.0004348783
#> [76] 0.0004539123 0.0004688737 0.0004785515 0.0004836363 0.0004867203
#> [81] 0.0004894326 0.0004920402 0.0004938790 0.0004939679 0.0004915013
#> [86] 0.0004861015 0.0004778602 0.0004672433 0.0004547678 0.0004407559
#> [91] 0.0004254074 0.0004088709
#> 
#> $SE_loghazard
#>  [1] 0.40543750 0.24826707 0.14315366 0.09378982 0.09031250 0.09802052
#>  [7] 0.10020644 0.09690515 0.09267049 0.09098529 0.09112544 0.09032418
#> [13] 0.08838373 0.08687533 0.08772633 0.09194764 0.09892862 0.10680319
#> [19] 0.11331657 0.11649595 0.11537883 0.11175247 0.10880134 0.10941833
#> [25] 0.11491230 0.12437381 0.13541850 0.14540330 0.15224260 0.15498501
#> [31] 0.15523397 0.15751592 0.16581058 0.18146588 0.20283833 0.22660276
#> [37] 0.24917759 0.26748722 0.27936202 0.28417416 0.28507230 0.28726922
#> [43] 0.29491645 0.30970145 0.33037935 0.35353593 0.37485252 0.39012134
#> [49] 0.39593948 0.39127790 0.38029923 0.36877390 0.36202226 0.36343465
#> [55] 0.37328141 0.38879016 0.40550109 0.41896821 0.42630956 0.42932111
#> [61] 0.43414179 0.44652168 0.46963172 0.50308230 0.54355103 0.58636409
#> [67] 0.62702400 0.66236850 0.69195910 0.72038382 0.75337194 0.79488095
#> [73] 0.84638388 0.90707419 0.97478897 1.04722492 1.12312982 1.20337950
#> [79] 1.29264372 1.39882960 1.52851450 1.68562356 1.87146828 2.08548657
#> [85] 2.32622226 2.59220233 2.88258497 3.19762092 3.53895558 3.90814616
#> [91] 4.30567477 4.73099657
#> 
#> $SE_log10hazard
#>  [1] 0.17607927 0.10782102 0.06217084 0.04073240 0.03922222 0.04256977
#>  [7] 0.04351911 0.04208537 0.04024628 0.03951441 0.03957528 0.03922729
#> [13] 0.03838457 0.03772948 0.03809906 0.03993235 0.04296415 0.04638404
#> [19] 0.04921276 0.05059355 0.05010839 0.04853348 0.04725182 0.04751978
#> [25] 0.04990578 0.05401486 0.05881151 0.06314785 0.06611812 0.06730913
#> [31] 0.06741726 0.06840829 0.07201062 0.07880963 0.08809157 0.09841233
#> [37] 0.10821645 0.11616822 0.12132538 0.12341527 0.12380533 0.12475944
#> [43] 0.12808059 0.13450163 0.14348193 0.15353870 0.16279638 0.16942754
#> [49] 0.17195433 0.16992983 0.16516186 0.16015647 0.15722427 0.15783767
#> [55] 0.16211406 0.16884942 0.17610689 0.18195558 0.18514389 0.18645179
#> [61] 0.18854538 0.19392190 0.20395846 0.21848587 0.23606121 0.25465469
#> [67] 0.27231306 0.28766298 0.30051402 0.31285872 0.32718527 0.34521241
#> [73] 0.36757985 0.39393731 0.42334547 0.45480400 0.48776908 0.52262108
#> [79] 0.56138804 0.60750398 0.66382541 0.73205701 0.81276835 0.90571531
#> [85] 1.01026549 1.12577917 1.25189075 1.38870912 1.53694888 1.69728631
#> [91] 1.86993079 2.05464570
#>