Get the Hazard Ratios with their Standard Errors

get_hr() takes as input the results of a model with covariates estimated by fit2ts or fit1ts and returns the estimated hazard ratios together with their standard errors.

Usage

get_hr(fitted_model)

Arguments

fitted_model

A list returned by the function fit2ts or fit1ts.

Value

A list with the following elements:

HR A vector of hazard ratios (calculated as \(\exp(\hat\beta)\)).

SE_HR A vector of Standard Errors for the hazard ratios calculated via the delta method.

beta A vector of the estimated \(\hat\beta\) coefficients.

SE_beta A vector of the Standard Errors for the beta coefficients.