Plot method for a haz2ts object.

plot.haz2ts() is the plot method for objects of class haz2ts. It produces plots of the fitted model with two time scales (see fit2ts()), either in the original (t,s) plane, while respecting the constraint imposed by the relation of the two time scales, or in the transformed (u,s) plane.

Usage

# S3 method for haz2ts
plot(
  x,
  plot_grid = NULL,
  which_plot = c("hazard", "covariates", "SE", "slices"),
  where_slices = NULL,
  direction = c(NULL, "u", "s"),
  plot_options = list(),
  ...
)

Arguments

x

The output of the function fit2ts. This is an object of class "haz2ts".

plot_grid

(optional) A list containing the parameters to build a new finer grid of intervals over u and s for plotting. This must be of the form: plot_grid = list(c(umin, umax, du), c(smin, smax, ds)), where umin, umax and smin, smax are the minimum and maximum values desired for the intervals over u and s respectively, and du, ds are distances between intervals over u and s respectively. Specifying a new denser grid is used to evaluate the B-spline bases used for estimation on such grid and plot the estimated surfaces with a greater level of details. If not specified, the plotting is done using the same B-splines bases as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

which_plot

The type of plot required. Can be one of "hazard" (default), "covariates", "SE" or "slices".

where_slices

A vector of values for the cutting points of the desired slices of the surface. If which_plot == "slices", please provide this argument.

direction

If which_plot == "slices", indicates the direction for cutting the surface. If u, then the surface will be cut at the selected values of u (indicated by where_slices), hence obtaining one-dimensional curves over s. If s, then the surface will be cut at the seleced values of s (indicated by where_slices), hence obtaining one-dimensional curves over u.

plot_options

A list with all possible options for any of the plots:

loghazard A Boolean. Default is FALSE. If FALSE the function returns a plot of the hazard surface, if TRUE the function returns a plot of the log-hazard surface.

log10hazard A Boolean. Default is FALSE. If TRUE, then a log_10 hazard surface is plotted.

cut_extrapolated A Boolean. Default is TRUE. Cuts away the extrapolated area of the (log-)hazard surface before plotting.

rectangular_grid A Boolean. Default is FALSE. If TRUE, a rectangular grid is used for plotting also in the (t,s)-plane as opposed to the grid of parallelograms used as default in the (t,s)-plane.

original A Boolean. Default is TRUE. Plot the (log-)hazard (and/or the SEs) in the (t,s)-plane. If FALSE, the (log-)hazard (and/or the SEs) will be plotted in the (u,s)-plane.

tmax The maximum value of t that should be plotted.

col_palette A function defining the color palette. The default palette is viridis::rev(plasma()).

n_shades The number of color shades to plot, default is 50.

breaks The vector of breaks for the color legend. If n_shades is provided, this should be of length n_shades + 1.

show_legend A Boolean. Default is TRUE. If FALSE no legend will be plotted, useful for multi-panel figures with common legend. Works only for plots on rectangular grid (i.e. transformed (u,s) plane)

main The title of the plot.

xlab The label of the first time axis (plotted on the x axis).

ylab The label of the second time axis (plotted on the y axis).

xlim A vector with two elements defining the limits of the time scale on the x axis.

ylim A vector with two elements defining the limits of the time scale on the y axis.

contour_lines A Boolean. Default is FALSE. If TRUE white contour lines are added to the surfaces.

contour_col The color for the contour lines. Default is white.

contour_cex The magnification to be used for the contour lines. Default is .8.

contour_nlev The number of contour levels desired. Default is 10.

cex_main The magnification to be used for the main title, default is 1.2 .

cex_lab The magnification to be used for the axis labels, default is 1 .

HR A Boolean. If TRUE the HRs with their CIs will be plotted. Default is FALSE (plot the beta with their CIs).

symmetric_CI A Boolean. Default is TRUE. If a plot of the HRs is required (HR == TRUE), then plot symmetrical Confidence Intervals, based on the SEs for the HRs calculated by delta method. If FALSE, then CIs are obtained by exponentiating the CIs for the betas.

confidence The level of confidence for the CIs. Default is .95 (alpha = 0.05).

col_beta The color for the plot of the covariates' effects.

pch The symbol for plotting the point estimates.

lwd The line width.

...

Further arguments to image.plot or image

Value

A plot of the fitted model.