Survival surface (cause-specific) and overall survival with two time scales

Computes cause-specific survival estimates over two time scales from a list of cause-specific cumulated hazard matrices. Each matrix of cause-specific survival is calculated as \(\exp(-CumHaz)\). Additionally, if the user provides more than one cause-specific cumulated hazard, the function computes the overall survival matrix, that contains the probability of not experiencing an event of any cause by time s and fixed time at entry u.

Usage

surv2ts(cumhaz = list(), cause = NULL)

Arguments

cumhaz

a list with all the cause-specific cumulated hazard matrices (minimum one element needs to be supplied). If more than one cause-specific cumulated hazard is provided, then they should all be matrices of the same dimension.

cause

is an optional vector of short names for the causes. It should be of the same length as the number of cause-specific cumulated hazards provided.

Value

a list with varying length. If only one cumulated hazard is provided as input, the function will calculate one two-dimensional survival surface. The output will be a list of one element. If K (with K>1) cumulated hazards are supplied, then the output will be a list with K+1 elements: * the first K elements will be the K cause-specific survival estimates; if a vector of short names with the causes is passed to the argument cause, then these will be used to name elements in the output; * OverSurv is the overall two-dimensional survival surface.