Plot of the covariates' effects

covariates_plot() produces a plot of the covariates' effects (\(\hat\beta\)) with confidence intervals, or of the Hazard Ratios (\(\exp(\hat\beta)\)) with confidence intervals.

Usage

covariates_plot(
  fitted_model,
  confidence_lev = 0.95,
  plot_options = list(),
  ...
)

Arguments

fitted_model

A list returned by the function fit2ts or fit1ts.

confidence_lev

The level of confidence for the CIs. Default is 0.95 (\(\alpha = 0.05\)).

plot_options

A list of options for the plot:

HR A Boolean. If TRUE the HRs with their CIs will be plotted. Default is FALSE (plot the beta with their CIs).

symmetric_ci A Boolean. Default is TRUE. If a plot of the HRs is required (HR == TRUE), then plot symmetrical Confidence Intervals, based on the SEs for the HRs calculated by delta method. If FALSE, then CIs are obtained by exponentiating the CIs for the betas.

main The title of the plot.

ylab The label for the y-axis.

ylim A vector with two elements defining the limits for the y-axis.

col_beta The color for the plot of the covariates' effects.

pch The symbol for plotting the point estimates.

cex_main The magnification factor for the main of the plot.

...

further arguments passed to plot()

Value

A plot of the covariates' effects. The different covariates are plotted on the x-axis, and on the y-axis the effects on the coefficient- or on the HR-scale are plotted. The main estimate is represented by a point and the CIs are added as vertical bars.