Get estimated (log-)hazard curve with 1 time scale

get_hazard_1d() takes as input the results of a model estimated by fit1ts and it returns the estimated smooth log-hazard and the smooth hazard together with their standard errors.

If the model includes covariates, then only the baseline (log-)hazard is returned. It is possible to provide values that define a new grid for plotting. If not specified, the plotting is done using the same B-splines basis as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Usage

get_hazard_1d(fitted_model, plot_grid = NULL)

Arguments

fitted_model

is an object of class "haz1ts", the output of the function fit1ts().

plot_grid

(optional) A named vector containing the parameters to build a new grid of intervals over s for plotting the estimated hazard on a finer grid. This must be of the form: plot_grid = c(smin, smax, ds), where smin, smax are the minimum and maximum values desired for the intervals over s, and ds is the distance between intervals over s. If not specified, the plotting is done using the same B-splines basis as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Value

A list with the following elements:

new_plot_grid A list of new specifications of the grid for plotting.

hazard A vector containing the estimated hazard.

loghazard A vector containing the estimated log-hazard.

log10hazard A vector containing the estimated log10-hazard.

SE_hazard A vector containing the estimated SEs for the hazard.

SE_loghazard A vector containing the estimated SEs for the log-hazard.

SE_log10hazard A vector containing the estimated SEs for the log10-hazard