Get estimated (log-)hazard values with 1 time scale

get_hazard_1d() takes as input the results of a model estimated by fit1ts and it returns the estimated values of the smooth log-hazard and the smooth hazard together with their standard errors.

If the model includes covariates, then only the baseline (log-)hazard is returned. It is possible to provide values that define a new grid for evaluation of the estimated hazard. If not specified, the hazard is evaluated on the same grid used for the binning of the data, and therefore the estimation of the model. The function will check if the parameters for the new grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Usage

get_hazard_1d(fitted_model, plot_grid = NULL)

Arguments

fitted_model

is an object of class "haz1ts", the output of the function fit1ts().

plot_grid

(optional) A named vector containing the parameters to build a new grid of intervals over s for plotting the estimated hazard on a finer grid. This must be of the form: plot_grid = c(smin, smax, ds), where smin, smax are the minimum and maximum values desired for the intervals over s, and ds is the distance between intervals over s. If not specified, the plotting is done using the same B-splines basis as for the estimation. The function will check if the parameters for the grid provided by the user are compatible with those originally used to construct the B-splines for estimating the model. If not, the grid will be adjusted accordingly and a warning will be returned.

Value

A list with the following elements:

new_plot_grid A list of parameters that specify the new grid, of the form list("ints", "smin", "smax", "ds")

hazard A vector containing the estimated hazard values.

loghazard A vector containing the estimated log-hazard values.

log10hazard A vector containing the estimated log10-hazard values.

SE_hazard A vector containing the estimated SEs for the hazard.

SE_loghazard A vector containing the estimated SEs for the log-hazard.

SE_log10hazard A vector containing the estimated SEs for the log10-hazard.

Examples

## preparing data - no covariates
dt1ts <- prepare_data(data = reccolon2ts,
                      s_in = "entrys",
                      s_out = "timesr",
                      events = "status",
                      ds = 180)

## fitting the model with fit1ts() - default options

mod1 <- fit1ts(dt1ts)
# Obtain 1d hazard
get_hazard_1d(mod1)
#> $new_plot_grid
#> $new_plot_grid$ints
#>  [1]    0  180  360  540  720  900 1080 1260 1440 1620 1800 1980 2160 2340 2520
#> [16] 2700 2880
#> 
#> $new_plot_grid$smin
#> [1] 0
#> 
#> $new_plot_grid$smax
#> [1] 2880
#> 
#> $new_plot_grid$ds
#> [1] 180
#> 
#> 
#> $hazard
#>               [,1]
#>  [1,] 0.0015531449
#>  [2,] 0.0018420923
#>  [3,] 0.0020660741
#>  [4,] 0.0021276863
#>  [5,] 0.0019876180
#>  [6,] 0.0016722017
#>  [7,] 0.0012900366
#>  [8,] 0.0009775307
#>  [9,] 0.0007626566
#> [10,] 0.0006230891
#> [11,] 0.0005248473
#> [12,] 0.0004479707
#> [13,] 0.0003840641
#> [14,] 0.0003291814
#> [15,] 0.0002815270
#> [16,] 0.0002404806
#> [17,] 0.0002053373
#> 
#> $loghazard
#>            [,1]
#>  [1,] -6.467473
#>  [2,] -6.296853
#>  [3,] -6.182105
#>  [4,] -6.152720
#>  [5,] -6.220818
#>  [6,] -6.393614
#>  [7,] -6.653085
#>  [8,] -6.930481
#>  [9,] -7.178703
#> [10,] -7.380821
#> [11,] -7.552403
#> [12,] -7.710783
#> [13,] -7.864701
#> [14,] -8.018902
#> [15,] -8.175282
#> [16,] -8.332871
#> [17,] -8.490857
#> 
#> $log10hazard
#>            [,1]
#>  [1,] -2.808788
#>  [2,] -2.734689
#>  [3,] -2.684854
#>  [4,] -2.672092
#>  [5,] -2.701667
#>  [6,] -2.776711
#>  [7,] -2.889398
#>  [8,] -3.009870
#>  [9,] -3.117671
#> [10,] -3.205450
#> [11,] -3.279967
#> [12,] -3.348750
#> [13,] -3.415596
#> [14,] -3.482565
#> [15,] -3.550480
#> [16,] -3.618920
#> [17,] -3.687532
#> 
#> $SE_hazard
#>               [,1]
#>  [1,] 0.0002526441
#>  [2,] 0.0002000306
#>  [3,] 0.0002039537
#>  [4,] 0.0002209619
#>  [5,] 0.0002334321
#>  [6,] 0.0002297991
#>  [7,] 0.0002100782
#>  [8,] 0.0001881904
#>  [9,] 0.0001742950
#> [10,] 0.0001724719
#> [11,] 0.0001826458
#> [12,] 0.0002025360
#> [13,] 0.0002289142
#> [14,] 0.0002572516
#> [15,] 0.0002835436
#> [16,] 0.0003053558
#> [17,] 0.0003214851
#> 
#> $SE_loghazard
#>             [,1]
#>  [1,] 0.16266613
#>  [2,] 0.10858879
#>  [3,] 0.09871559
#>  [4,] 0.10385077
#>  [5,] 0.11744312
#>  [6,] 0.13742310
#>  [7,] 0.16284667
#>  [8,] 0.19251606
#>  [9,] 0.22853662
#> [10,] 0.27680127
#> [11,] 0.34799800
#> [12,] 0.45211885
#> [13,] 0.59603121
#> [14,] 0.78148889
#> [15,] 1.00716301
#> [16,] 1.26977289
#> [17,] 1.56564416
#> 
#> $SE_log10hazard
#>             [,1]
#>  [1,] 0.07064500
#>  [2,] 0.04715951
#>  [3,] 0.04287163
#>  [4,] 0.04510182
#>  [5,] 0.05100490
#>  [6,] 0.05968209
#>  [7,] 0.07072341
#>  [8,] 0.08360866
#>  [9,] 0.09925219
#> [10,] 0.12021326
#> [11,] 0.15113361
#> [12,] 0.19635272
#> [13,] 0.25885306
#> [14,] 0.33939631
#> [15,] 0.43740534
#> [16,] 0.55145536
#> [17,] 0.67995062
#>